pyproject.toml
[tool.poetry]
name = "select-prefetch-example"
version = "0.1.0"
description = ""
authors = ["Kader Miyanyedi <kadermiyanyedi@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
django-debug-toolbar = "^3.2.4"
django-extensions = "^3.1.5"
bpython = "^0.22.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
